Rename service functions
