+++ src/compiler/typescript_to_rust/azle_generate/src/body/ic_object/functions/mod.rs
@@ -13,7 +13,6 @@ mod candid_encode;
 mod canister_balance;
 mod canister_balance128;
 mod clear_timer;
-mod cross_canister_calls;
 mod data_certificate;
 mod id;
 mod method_name;
@@ -31,6 +30,7 @@ mod reject_code;
 mod reject_message;
 mod reply;
 mod reply_raw;
+mod service;
 mod set_certified_data;
 mod set_timer;
 mod set_timer_interval;
@@ -57,7 +57,6 @@ pub fn generate(
     let arg_data_raw_size = arg_data_raw_size::generate();
     let call_raw = call_raw::generate();
     let call_raw128 = call_raw128::generate();
-    let cross_canister_calls = cross_canister_calls::generate(services);
     let caller = caller::generate();
     let candid_decode = candid_decode::generate();
     let candid_encode = candid_encode::generate();
@@ -81,6 +80,7 @@ pub fn generate(
     let reject_message = reject_message::generate();
     let reply = reply::generate(query_and_update_methods);
     let reply_raw = reply_raw::generate();
+    let service_functions = service::generate(services);
     let set_certified_data = set_certified_data::generate();
     let set_timer = set_timer::generate();
     let set_timer_interval = set_timer_interval::generate();
@@ -88,6 +88,7 @@ pub fn generate(
     let stable64_read = stable64_read::generate();
     let stable64_size = stable64_size::generate();
     let stable64_write = stable64_write::generate();
+    let stable_b_tree_map = stable_b_tree_map::generate(stable_b_tree_map_nodes);
     let stable_bytes = stable_bytes::generate();
     let stable_grow = stable_grow::generate();
     let stable_read = stable_read::generate();
@@ -95,7 +96,6 @@ pub fn generate(
     let stable_write = stable_write::generate();
     let time = time::generate();
     let trap = trap::generate();
-    let stable_b_tree_map = stable_b_tree_map::generate(stable_b_tree_map_nodes);
 
     quote::quote! {
         #accept_message
@@ -103,7 +103,6 @@ pub fn generate(
         #arg_data_raw_size
         #call_raw
         #call_raw128
-        #(#cross_canister_calls)*
         #caller
         #candid_decode
         #candid_encode
@@ -127,14 +126,15 @@ pub fn generate(
         #reject_message
         #reply
         #reply_raw
+        #(#service_functions)*
         #set_certified_data
         #set_timer
         #set_timer_interval
-        #stable_b_tree_map
         #stable64_grow
         #stable64_read
         #stable64_size
         #stable64_write
+        #stable_b_tree_map
         #stable_bytes
         #stable_grow
         #stable_read
